requirements.txt
numpy==1.23.5
pandas==1.5.3
scikit_learn==1.2.1
scipy==1.9.3
sklearn_relief==1.0.0b2
